+++ src/lib.rs
@@ -26,10 +26,10 @@
 //!    db.the_awnser().set(&42).unwrap();
 //!	   assert_eq!(42u8, db.the_awnser().get().unwrap());
 //!
-//!    db.primes().push(2).unwrap();
-//!    db.primes().push(3).unwrap();
-//!    db.primes().push(5).unwrap();
-//!	   db.primes().push(7).unwrap();
+//!    db.primes().push(&2).unwrap();
+//!    db.primes().push(&3).unwrap();
+//!    db.primes().push(&5).unwrap();
+//!	   db.primes().push(&7).unwrap();
 //!	   assert_eq!(Some(7), db.primes().pop().unwrap());
 //!
 //!	   assert_eq!(String::from("42"), db.the_result().get().unwrap());
@@ -81,6 +81,7 @@ pub use dbstruct_derive::*;
 pub mod stores;
 pub mod traits;
 pub use traits::{ByteStore, DataStore};
+pub use traits::{TryExtend};
 pub mod wrapper;
 
 pub use sled;