+++ CHANGELOG.md
@@ -7,14 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.2.1] - 2023-08-12
+
+## [0.4.0] - 2023-08-17
+
+### Added
+ - adds TryExtend that mirrors the std's Extend trait
+
+### Changed
+ - wrapper::Vec::push now takes a reference to the item to be pushed
+
+## [0.3.0] - 2023-08-12
 
 ### Added
  - adds iterator methods to the map wrapper
  - adds BTreeMap backend
 
-### Changed
- - deprecates HashMap test backend
+### Deprecated
+ - HashMap test backend (use new BTreeMap backend)
 
 ## [0.2.1] - 2023-08-02